[feat] 은퇴시뮬레이션 dto 추가, 파일 위치 오류 수정
