+++ app/src/main/java/com/moneymate/moneymate/data/dto/mypage/response/RetireResultResponse.kt
@@ -1,0 +1,16 @@
+package com.moneymate.moneymate.data.dto.mypage.response
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RetireResultResponse(
+    val assets: List<Asset>
+)
+
+@Serializable
+data class Asset(
+    val age: Int,
+    val asset: Long,
+    val income: Long,
+    val expense: Long
+)
