+++ app/src/main/java/com/moneymate/moneymate/ui/manage/screen/RetireInputScreen.kt
@@ -1,4 +1,4 @@
-package com.moneymate.moneymate.ui.mypage.screen
+package com.moneymate.moneymate.ui.manage.screen
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -13,8 +13,6 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Icon
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
