+++ app/src/main/java/com/moneymate/moneymate/ui/manage/screen/RetireResultScreen.kt
@@ -1,9 +1,8 @@
-package com.moneymate.moneymate.ui.mypage.screen
+package com.moneymate.moneymate.ui.manage.screen
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
